Add a `collectGithubRepoFiles` tool
